refactor(parallel_executor): split 2946-line file into module dir
Decomposes src/parallel_executor.rs into src/parallel_executor/ (10 files)
by responsibility. Pure code-move: behavior byte-identical, only visibility
and import glue added.

- error.rs      ParallelExecutorError (23)
- config.rs     ExecutorConfig + presets + effective_* accessors (106)
- summary.rs    ExecutionSummary + metric DTOs + impl (305)
- pool.rs       ConnectionPool (144)
- batch.rs      BatchProcessor (150)
- monitor.rs    PerformanceMonitor + MemoryTracker + ThroughputCalculator (183)
- lua_bridge.rs SerializedValue/Function + LuaContextFactory + ExecutionResult
                + parallel_executor_constructor + create_global_executor_interface
                + impl UserData (489)
- validation.rs validate_config + validate_config_combinations, converted from
                associated fns of ParallelExecutor to pub(crate) free fns (166)
- mod.rs        ParallelExecutor struct + impl (new/map/configure/accessors) +
                GLOBAL_EXECUTOR + global_executor + init_global_executor +
                submodule decls + re-exports (431)
- tests.rs      #[cfg(test)] unit tests, moved wholesale (990)

validate_config was a private associated fn used in new(), the UserData
'validate_config' method, and tests; now a free fn re-exported crate-internally.
All call sites converted (Self::/ParallelExecutor:: -> bare call). Public API
preserved via pub use (items were already pub in the monolith).
